requirements.txt
chromadb==1.0.20
pymupdf==1.26.4
sentence-transformers==5.1.0
python-dotenv==1.1.1
google-generativeai==0.8.5
faster-whisper==1.0.3